Move test_utils to utilties
